# api/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "Mobius",
    "version": "0.1.0"
  },
  "paths": {
    "/api/agents": {
      "get": {
        "description": "ListActiveAgents returns the open sessions of all organizations in one group for each Role, with the slot counts and the limits of the config.",
        "operationId": "ListActiveAgents",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ActiveAgents"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/{id}/transcript": {
      "get": {
        "description": "GetTranscript returns the Transcript of a session, the oldest row first.",
        "operationId": "GetTranscript",
        "parameters": [
          {
            "description": "ID is the id of the session",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Array_TranscriptLine"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "agents"
        ]
      }
    },
    "/api/chat": {
      "get": {
        "description": "GetChat returns the Lead chat of a Workstream, or the Triager chat of an organization.",
        "operationId": "GetChat",
        "parameters": [
          {
            "description": "Organization is the owner of the repository, or the organization of the Triager chat",
            "in": "query",
            "name": "organization",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Repository is the repository of the Workstream as \"owner/name\". It is empty for the Triager chat",
            "in": "query",
            "name": "repository",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Workstream is the number of the Workstream issue. It is 0 for the Triager chat",
            "in": "query",
            "name": "workstream",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Chat"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/messages": {
      "post": {
        "description": "SendChat adds a message of the Owner with its images to the chat, and gives it to the Lead or to the Triager. Mobius starts the agent when none runs. It returns 400 when the request is larger than 4 images of 5 MB with the text. It returns 409 when the organization has no repository of Mobius, or while Mobius restarts for an upgrade.",
        "operationId": "SendChat",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SendChatBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/messages/{id}/images/{position}": {
      "get": {
        "description": "GetChatImage returns an image of a message. It returns 404 when the message has no such image.",
        "operationId": "GetChatImage",
        "parameters": [
          {
            "description": "ID is the id of the message",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Position is the position of the image in the message, from 0",
            "in": "path",
            "name": "position",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/gif": {
                "schema": {
                  "contentMediaType": "image/gif",
                  "type": "string"
                }
              },
              "image/jpeg": {
                "schema": {
                  "contentMediaType": "image/jpeg",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "contentMediaType": "image/png",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "contentMediaType": "image/webp",
                  "type": "string"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/seen": {
      "post": {
        "description": "SeeChat records that the Owner saw the messages of the chat up to a message. An earlier message than the last seen message changes nothing.",
        "operationId": "SeeChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeeChatBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/stop": {
      "post": {
        "description": "StopChat stops the turn of the agent that answers a message of the Owner. A turn for a Lead event goes on. When the agent waits for its slot, its first message of the Owner gets no turn.",
        "operationId": "StopChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopChatBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "chat"
        ]
      }
    },
    "/api/checkup": {
      "get": {
        "description": "GetCheckup returns the status of the App permissions and of each Mobius label in each repository of the organization.",
        "operationId": "GetCheckup",
        "parameters": [
          {
            "description": "Organization is the owner of the repositories",
            "in": "query",
            "name": "organization",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Checkup"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "checkup"
        ]
      }
    },
    "/api/checkup/fix": {
      "post": {
        "description": "FixLabels creates the missing Mobius labels and sets the fixed color of each Mobius label with a different color, in each repository of the organization. The description of an existing label stays.",
        "operationId": "FixLabels",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixLabelsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "checkup"
        ]
      }
    },
    "/api/checkup/tools": {
      "get": {
        "description": "GetCheckupTools returns the path and the version of each program that Mobius runs for its agents and its work.",
        "operationId": "GetCheckupTools",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Array_ToolCheck"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "checkup"
        ]
      }
    },
    "/api/devices": {
      "get": {
        "description": "ListDevices returns the device logins.",
        "operationId": "ListDevices",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Devices"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "devices"
        ]
      }
    },
    "/api/devices/{id}": {
      "delete": {
        "description": "Logout deletes a device login.",
        "operationId": "Logout",
        "parameters": [
          {
            "description": "ID is the id of the device login",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "devices"
        ]
      }
    },
    "/api/drain": {
      "get": {
        "description": "GetDrain returns the state of the drain.",
        "operationId": "GetDrain",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Drain"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "upgrade"
        ]
      },
      "post": {
        "description": "StartDrain holds each new Worker in the queue, and returns when no agent runs or when the Owner cancels the drain.",
        "operationId": "StartDrain",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DrainEnd"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "upgrade"
        ]
      },
      "delete": {
        "description": "CancelDrain ends the drain, so the held Workers start. It does nothing while the upgrade replaces the program.",
        "operationId": "CancelDrain",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "upgrade"
        ]
      }
    },
    "/api/events": {
      "get": {
        "description": "StreamEvents sends the latest activities, and then each new activity, each change of a session, each new or changed Transcript row, each change of the drain, each start and failure of an upgrade, each change of the Workstream list, each new or longer chat message, each new unread count, each change of the state of a chat, each new or dismissed Inbox item, and each Workstream that the Triager chat created. When the request has Last-Event-ID, it sends the activities after that id in place of the latest activities. When the client does not read the session changes fast enough, the stream ends.",
        "operationId": "StreamEvents",
        "parameters": [
          {
            "description": "LastEventID is the id of the last activity that the client got",
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "$ref": "#/components/schemas/LiveEvents"
                }
              }
            },
            "description": "Event stream"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "events"
        ]
      }
    },
    "/api/github/apps": {
      "get": {
        "description": "ListGitHubApps returns the Mobius Apps with the repositories of their installations.",
        "operationId": "ListGitHubApps",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Array_GitHubApp"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "github"
        ]
      }
    },
    "/api/github/manifest": {
      "post": {
        "description": "CreateManifestForm returns the form that creates a Mobius App in a GitHub account. After the form, GitHub sends the browser to GET /api/github/manifest-callback.",
        "operationId": "CreateManifestForm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManifestFormBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ManifestForm"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "github"
        ]
      }
    },
...
    "/api/inbox": {
      "get": {
        "description": "ListInbox returns the Inbox items that the Owner did not dismiss, the oldest first.",
        "operationId": "ListInbox",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Array_InboxItem"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "mobius_session": []
          }
        ],
        "tags": [
          "inbox"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_ActiveAgents": {
        "type": "object",
        "title": "Envelope_ActiveAgents",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ActiveAgents",
            "description": "Data is the payload of the response"
          }
        },
        "required": [
          "data"
        ],
        "description": "Envelope is the body of each success response."
      },
...
      "ErrorResponse": {
        "type": "object",
        "title": "ErrorResponse",
        "properties": {
          "details": {
            "type": "object",
            "description": "Additional error details",
            "additionalProperties": {}
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": [
          "error"
        ],
        "description": "Generic error response structure"
      },
...
    },
    "securitySchemes": {
      "mobius_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mobius_session"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
...
      "UnprocessableEntity": {
        "description": "Unprocessable Entity - Request body could not be parsed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}